// src/global/enums/EUserGlobal.ts
/*
 * @Author: Always
 * @LastEditors: Always
 * @Date: 2020-06-23 14:41:15
 * @LastEditTime: 2020-06-23 14:46:30
 * @FilePath: /koala-background-server/src/frontend/enums/EUserGlobal.ts
 */

// 用户性别
export enum EUserGender {
  UNKONWN = 0, // 未知
  MAN = 1, // 男性
  WOMAN = 2, // 女性
}

// 用户所用语言
export enum EUserLanguage {
  EN = 'en', // 英文
  ZH_CN = 'zh_CN', // 简体中文
  ZH_TW = 'zh_TW', // 繁体中文
}